npm v5.0.0 and v5.0.1 package-lock.json bug demo
This demo reproduces a couple of surprising scenarios arising from the
use of the `package-lock.json` file in npm v5.0.0 and v5.0.1. The
behavior described below is reproducible with both npm versions, and
appears related to the fact that `devDependency` packages whose own
dependencies have a version spec of `latest` are not handled properly
depending on the scenario. This demo uses `live-server` to illustrate
the phenomena.